Use ora to display current state

+++ bin/cli.js
@@ -5,6 +5,7 @@ const util = require('util');
 const exec = util.promisify(child_process.exec);
 
 const yargs = require('yargs');
+const ora = require('ora');
 const replaceInFiles = require('replace-in-files');
 
 const argv = yargs
@@ -25,21 +26,44 @@ const options = {
         `${name}/README.md`,
         `${name}/package.json`,
         `${name}/package-lock.json`,
-        `${name}/cli.js`,
+        `${name}/bin/cli.js`,
     ],
     from: /create-ts-project/g,
     to: name
 };
 
+const state = ora();
+state.start();
+
+async function isFileExist(name) {
+    try {
+        await exec(`stat ${name}`);
+        return true;
+    } catch (ignore) {
+        // console.log(ignore);
+        return false;
+    }
+}
+
 (async () => {
-    console.log(`Create project: ${name}`);
-    // Fetch github.com/piecioshka/create-ts-project
-    await exec(`wget ${package} -O create-ts-project.zip`);
-    await exec(`unzip create-ts-project.zip`);
-    await exec(`mv create-ts-project-master ${name}`);
-    await exec(`rm -rf create-ts-project.zip`);
-    // Replace all "create-ts-project" by "NAME"
-    await replaceInFiles(options);
-    // Git setup & commit
-    await exec(`cd ${name} && git init && git commit -am "Generate project"`);
+    state.info(`Create project: ${name}`);
+    try {
+        const isDirectoryExist = await isFileExist(name);
+        if (isDirectoryExist) {
+            throw new Error('Directory exist');
+        }
+        // Fetch github.com/piecioshka/create-ts-project
+        await exec(`wget ${package} -O create-ts-project.zip`);
+        await exec(`unzip create-ts-project.zip`);
+        await exec(`mv create-ts-project-master ${name}`);
+        await exec(`rm -rf create-ts-project.zip`);
+        // Replace all "create-ts-project" by "NAME"
+        await replaceInFiles(options);
+        // Git setup & commit
+        await exec(`cd ${name} && git init && git add . && git commit -am "Generate project"`);
+        state.succeed('Project created');
+        state.stop();
+    } catch (reason) {
+        state.fail(`Project does not created properly: ${reason.message}`);
+    }
 })();
